AppSentinel/AppScore/aws/lambda_functions/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.11
coverage==7.3.2
boto3==1.34.7
botocore==1.34.7
psycopg2-binary==2.9.9
